+++ src/base/net/stratum/DaemonClient.cpp
@@ -245,6 +245,10 @@ void xmrig::DaemonClient::setPool(const Pool &pool)
     if (!m_coin.isValid() && pool.algorithm() == Algorithm::RX_XEQ) {
         m_coin = Coin::XEQ;
     }
+
+    if (!m_coin.isValid() && pool.algorithm() == Algorithm::RX_VRL) {
+        m_coin = Coin::VRL;
+    }
 }