+++ src/crypto/randomx/randomx.h
@@ -150,6 +150,7 @@ struct RandomX_ConfigurationSafex : public RandomX_ConfigurationBase { RandomX_C
 struct RandomX_ConfigurationYada : public RandomX_ConfigurationBase { RandomX_ConfigurationYada(); };
 struct RandomX_ConfigurationTuske : public RandomX_ConfigurationBase { RandomX_ConfigurationTuske(); };
 struct RandomX_ConfigurationEquilibria : public RandomX_ConfigurationBase { RandomX_ConfigurationEquilibria(); };
+struct RandomX_ConfigurationVirel : public RandomX_ConfigurationBase { RandomX_ConfigurationVirel(); };
 
 extern RandomX_ConfigurationMonero RandomX_MoneroConfig;
 extern RandomX_ConfigurationWownero RandomX_WowneroConfig;
@@ -159,6 +160,7 @@ extern RandomX_ConfigurationSafex RandomX_SafexConfig;
 extern RandomX_ConfigurationYada RandomX_YadaConfig;
 extern RandomX_ConfigurationTuske RandomX_TuskeConfig;
 extern RandomX_ConfigurationEquilibria RandomX_EquilibriaConfig;
+extern RandomX_ConfigurationVirel RandomX_VirelConfig;
 
 extern RandomX_ConfigurationBase RandomX_CurrentConfig;